assessment task

+++ database/migrations/2022_05_16_143445_create_orders_table.php
@@ -18,10 +18,9 @@ return new class extends Migration
             $table->id();
             $table->foreignId('merchant_id')->constrained();
             $table->foreignId('affiliate_id')->nullable()->constrained();
-            // TODO: Replace floats with the correct data types (very similar to affiliates table)
             //We should user double here to get accuracy in sub total and commission
             $table->double('subtotal');
-            $table->double('commission_owed')->default(0.00);
+            $table->float('commission_owed')->default(0.00);
             $table->string('payout_status')->default(Order::STATUS_UNPAID);
             $table->string('discount_code')->nullable();
             $table->timestamps();
